Shader templates are now loaded dynamically from intellij template system
