+++ resources/fileTemplates/internal/GLSL Fragment Shader.frag.ft
