+++ resources/fileTemplates/internal/GLSL Vertex Shader.vert.ft
